Move ruma-identifiers validation logic into a new crate
