+++ ruma-identifiers-validation/src/error.rs
