+++ ruma-identifiers-validation/src/key_algorithms.rs
