+++ ruma-identifiers-validation/src/room_version_id.rs
@@ -1,0 +1,14 @@
+use crate::Error;
+
+/// Room version identifiers cannot be more than 32 code points.
+const MAX_CODE_POINTS: usize = 32;
+
+pub fn validate(s: &str) -> Result<(), Error> {
+    if s.is_empty() {
+        Err(Error::MinimumLengthNotSatisfied)
+    } else if s.chars().count() > MAX_CODE_POINTS {
+        Err(Error::MaximumLengthExceeded)
+    } else {
+        Ok(())
+    }
+}
