+++ ruma-identifiers/src/server_name.rs
@@ -6,7 +6,9 @@ use std::{
     mem,
 };
 
-use crate::error::Error;
+use ruma_identifiers_validation::server_name::validate;
+
+use crate::Error;
 
 /// A Matrix-spec compliant server name.
 #[repr(transparent)]
@@ -59,51 +61,6 @@ impl From<&ServerName> for Box<ServerName> {
     }
 }
 
-pub(crate) fn validate(server_name: &str) -> Result<(), Error> {
-    use std::net::Ipv6Addr;
-
-    if server_name.is_empty() {
-        return Err(Error::InvalidServerName);
-    }
-
-    let end_of_host = if server_name.starts_with('[') {
-        let end_of_ipv6 = match server_name.find(']') {
-            Some(idx) => idx,
-            None => return Err(Error::InvalidServerName),
-        };
-
-        if server_name[1..end_of_ipv6].parse::<Ipv6Addr>().is_err() {
-            return Err(Error::InvalidServerName);
-        }
-
-        end_of_ipv6 + 1
-    } else {
-        let end_of_host = server_name.find(':').unwrap_or_else(|| server_name.len());
-
-        if server_name[..end_of_host]
-            .bytes()
-            .any(|byte| !(byte.is_ascii_alphanumeric() || byte == b'-' || byte == b'.'))
-        {
-            return Err(Error::InvalidServerName);
-        }
-
-        end_of_host
-    };
-
-    if server_name.len() != end_of_host
-        && (
-            // hostname is followed by something other than ":port"
-            server_name.as_bytes()[end_of_host] != b':'
-            // the remaining characters after ':' are not a valid port
-            || server_name[end_of_host + 1..].parse::<u16>().is_err()
-        )
-    {
-        Err(Error::InvalidServerName)
-    } else {
-        Ok(())
-    }
-}
-
 fn try_from<S>(server_name: S) -> Result<Box<ServerName>, Error>
 where
     S: AsRef<str> + Into<Box<str>>,
@@ -140,7 +97,7 @@ impl<'a> TryFrom<&'a str> for &'a ServerName {
 }
 
 impl TryFrom<&str> for Box<ServerName> {
-    type Error = crate::error::Error;
+    type Error = crate::Error;
 
     fn try_from(s: &str) -> Result<Self, Self::Error> {
         try_from(s)
@@ -148,7 +105,7 @@ impl TryFrom<&str> for Box<ServerName> {
 }
 
 impl TryFrom<String> for Box<ServerName> {
-    type Error = crate::error::Error;
+    type Error = crate::Error;
 
     fn try_from(s: String) -> Result<Self, Self::Error> {
         try_from(s)
