+++ ruma-identifiers/src/user_id.rs
@@ -2,7 +2,7 @@
 
 use std::{convert::TryFrom, num::NonZeroU8};
 
-use crate::{error::Error, parse_id, ServerName};
+use crate::{Error, ServerName};
 
 /// A Matrix user ID.
 ///
@@ -95,41 +95,14 @@ fn try_from<S>(user_id: S) -> Result<UserId, Error>
 where
     S: AsRef<str> + Into<Box<str>>,
 {
-    let user_id_str = user_id.as_ref();
-
-    let colon_idx = parse_id(user_id_str, &['@'])?;
-    let localpart = &user_id_str[1..colon_idx.get() as usize];
-
-    let is_historical = localpart_is_fully_comforming(localpart)?;
-
+    let (colon_idx, is_historical) =
+        ruma_identifiers_validation::user_id::validate(user_id.as_ref())?;
     Ok(UserId { full_id: user_id.into(), colon_idx, is_historical: !is_historical })
 }
 
 common_impls!(UserId, try_from, "a Matrix user ID");
 
-/// Check whether the given user id localpart is valid and fully conforming
-///
-/// Returns an `Err` for invalid user ID localparts, `Ok(false)` for historical user ID localparts
-/// and `Ok(true)` for fully conforming user ID localparts.
-pub fn localpart_is_fully_comforming(localpart: &str) -> Result<bool, Error> {
-    if localpart.is_empty() {
-        return Err(Error::InvalidLocalPart);
-    }
-
-    // See https://matrix.org/docs/spec/appendices#user-identifiers
-    let is_fully_conforming = localpart
-        .bytes()
-        .all(|b| matches!(b, b'0'..=b'9' | b'a'..=b'z' | b'-' | b'.' | b'=' | b'_' | b'/'));
-
-    // If it's not fully conforming, check if it contains characters that are also disallowed
-    // for historical user IDs. If there are, return an error.
-    // See https://matrix.org/docs/spec/appendices#historical-user-ids
-    if !is_fully_conforming && localpart.bytes().any(|b| b < 0x21 || b == b':' || b > 0x7E) {
-        Err(Error::InvalidCharacters)
-    } else {
-        Ok(is_fully_conforming)
-    }
-}
+pub use ruma_identifiers_validation::user_id::localpart_is_fully_comforming;
 
 #[cfg(test)]
 mod tests {
@@ -139,7 +112,7 @@ mod tests {
     use serde_json::{from_str, to_string};
 
     use super::UserId;
-    use crate::{error::Error, ServerName};
+    use crate::{Error, ServerName};
 
     #[test]
     fn valid_user_id_from_str() {
